requirements.txt
torch==1.13
torchvision==0.14
scikit-learn==1.2.2
pandas==1.5.3
scipy==1.10.1
tqdm==4.65.0
matplotlib==3.7.1
pyyaml==6.0
hydra-core==1.3.2
torchsummary==1.5.1
hydra-optuna-sweeper==1.2.0
uvicorn[standard]==0.14.0
gunicorn==20.1.0
fastapi
scikit-image
python-multipart==0.0.5
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
altair < 5
ipykernel==6.9.2
# - mlflow==1.22.0
mlflow==2.5.0
pylint==2.6.0
optuna==2.10.0
protobuf==3.20.0
numpy==1.21.6
# - alembic==1.11.1
plotly==5.15.0